+++ mongsom/dev/src/main/java/com/mongsom/dev/dto/cart/reqDto/CartChangeCheckReqDto.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class CartChangeCheckReqDto {
     
     @NotNull(message = "사용자 코드는 필수입니다")
-    private Integer userCode;
+    private Long userCode;
     
     @NotNull(message = "상품 ID는 필수입니다")
     private Integer productId;
